% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_lake_wiki}
\alias{get_lake_wiki}
\title{get_lake_wiki}
\usage{
get_lake_wiki(lake_name)
}
\arguments{
\item{lake_name}{character}
}
\description{
get_lake_wiki
}
\examples{
\dontrun{
get_lake_wiki("Lake Nipigon")
}
}

