% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{lake_clean}
\alias{lake_clean}
\title{Clean output of lake_wiki}
\usage{
lake_clean(dt)
}
\arguments{
\item{dt}{output of the lake_wiki function}
}
\description{
Currently the only operation is to standardize the units of numeric fields.
See the output units with the unit_key_ function.
}
\examples{
\dontrun{
dt <- lake_wiki(c("Lake Mendota","Flagstaff Lake (Maine)"))
dt_clean <- lake_clean(dt)

dt <- lake_wiki(c("Lake Mendota","Trout Lake (Wisconsin)"))
dt_clean <- lake_clean(dt)
}
}
