% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{lake_wiki_browser}
\alias{lake_wiki_browser}
\title{lake_wiki_browser}
\usage{
lake_wiki_browser(lake_wiki_obj = NA, lake_names = NA)
}
\arguments{
\item{lake_wiki_obj}{data.frame output of lake_wiki}

\item{lake_names}{fallback character vector of lake names}
}
\description{
lake_wiki_browser
}
\examples{
\dontrun{
lake_wiki_browser(lake_names = "Lake Mendota")
lake_wiki_browser(lake_names = c("Lake Mendota", "Lake Champlain"))
lake_wiki_browser(lake_wiki(c("Lake Mendota", "Lake Champlain")))
}
}
