% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikilake-package.R
\docType{data}
\name{milakes}
\alias{milakes}
\title{Michigan Lakes}
\format{A data frame with 48 columns and 177 rows:
\itemize{
          \item{Name: }{lake name}
          \item{Location: }{location description}
          \item{Primary inflows: }{rivers and streams}
          \item{Basin countries: }{countries}
          \item{Surface area: }{hectares}
          \item{Max. depth: }{meters}
          \item{Surface elevation: }{meters}
          \item{Lat: }{decimal degrees}
          \item{Lon: }{decimal degrees}
          \item{Primary outflows: }{rivers and streams}
          \item{Average depth: }{meters}
          \item{Max. length: }{meters}
          \item{Max. width: }{meters}
          }}
\description{
Metadata of Michigan lakes scraped from Wikipedia.
}
\keyword{datasets}

