% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_unit_brackets}
\alias{parse_unit_brackets}
\title{Parse string representation of units package quantities}
\usage{
parse_unit_brackets(x, target_unit = NA)
}
\arguments{
\item{x}{character string with unit in brackets}

\item{target_unit}{target unit to convert to. optional}
}
\description{
Parse string representation of units package quantities
}
\examples{
x <- "1 [m]"
x <- "8.5 [m]"
parse_unit_brackets(x, "feet")
}
