% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikilake-package.R
\docType{package}
\name{wikilake-package}
\alias{wikilake}
\alias{wikilake-package}
\title{Scrape Wikipedia lakes metadata}
\description{
Scrape Wikipedia lakes metadata
}
\author{
\email{stachel2@msu.edu}
}

